python/django/requirements.txt
djangorestframework
psycopg2
